Add pint hook to pre-commit, ref gh-actions
Laravel's Pint will now be part of the pre-commit hooks used for the CI
pipeline.

Since this hook depends on a docker image, the project will stop using
pre-commit.ci (does not support it) and instead run a custom workflow.

The hook runs fine on local machine but for unkown reasons it fails on
Github's CI, therefore it runs as a separate step and it apparently
works there (who knows why).