CI: Use PostgreSQL, npm for tests
Feature tests are very likely to require a running database, therefore a
PostgreSQL service has been added to the tests job.

.env.testing was also removed since it seems like it is not a good
practice and may be annoying for devs that want to have a custom one for
their setup.

Finally, npm dependencies are installed and resources built as required
by a full on deployment. Migrations are run as well, this ensures that
tests run in an environment as close as possible to a production one.