IMP: fallback to last session if not provide on measurements
Previously creating new measurements always required the session id to
be specified. This turned out to be repetitive since boards generally
run a single session at a time and measurements were always taken for
the current session in use (last one to be created).

session_id is herefoe no longer required. The previous API usage is
still valid as well.